requirements.txt
Django==4.2
djangorestframework==3.14.0
django-ckeditor==6.7.0
redis==4.6.0
googletrans==4.0.0-rc1